@@ -1,60 +1,30 @@
 {
-  "name": "anythink-market-front",
+  "name": "mongo-chatbot-ui",
   "version": "0.1.0",
-  "engines": {
-    "node": "^16"
-  },
   "private": true,
-  "devDependencies": {
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
-    "core-js": "^3.25.1",
-    "enzyme": "^3.11.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-react": "^7.26.1",
-    "prettier": "2.4.1",
-    "react-test-renderer": "^17.0.2",
-    "redux-mock-store": "^1.5.4"
-  },
   "dependencies": {
-    "@babel/core": "^7.18.13",
-    "@babel/plugin-proposal-private-property-in-object": "^7.18.6",
-    "@babel/plugin-syntax-flow": "^7.18.6",
-    "@babel/plugin-transform-react-jsx": "^7.18.10",
-    "bootstrap": "^4.6.0",
-    "bootstrap-icons": "^1.7.1",
-    "history": "^4.6.3",
-    "jquery": "^3.6.1",
-    "marked": "^0.3.6",
-    "postcss": "^8.4.16",
-    "prop-types": "^15.5.10",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^5.0.7",
-    "react-router-dom": "^6.9.0",
-    "react-scripts": "^5.0.1",
-    "redux": "^3.6.0",
-    "redux-devtools-extension": "^2.13.2",
-    "sass": "^1.45.0",
-    "superagent": "^3.8.2",
-    "superagent-promise": "^1.1.0",
-    "typescript": "^4.8.2"
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "mongodb-chatbot-ui": "^0.1.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "5.0.1",
+    "web-vitals": "^2.1.4"
+  },
+  "devDependencies": {
+      "@types/node": "^20.10.5"
   },
   "scripts": {
-    "start": "REACT_APP_WILCO_ID=${WILCO_ID:-\"$(cat ../.wilco)\"} react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "format": "yarn prettier --write .",
-    "lint": "yarn eslint . && yarn prettier --check ."
+    "dev": "yarn build --watch & yarn start"
   },
   "eslintConfig": {
     "extends": [
       "react-app",
-      "eslint:recommended"
-    ],
-    "rules": {
-      "no-var": "error"
-    }
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -67,8 +37,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "resolutions": {
-    "autoprefixer": "10.4.5"
   }
 }
