@@ -1,18 +1,17 @@
-import "./custom.scss";
-import ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import React from "react";
-import { store } from "./store";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import App from './App';
+import reportWebVitals from './reportWebVitals';
 
-import App from "./components/App";
-import { BrowserRouter } from "react-router-dom";
-
-ReactDOM.render(
-  <Provider store={store}>
-    <BrowserRouter basename="/">
-      <App />
-    </BrowserRouter>
-  </Provider>,
-
-  document.getElementById("root")
+const root = ReactDOM.createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
 );
+
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
