@@ -11,7 +11,7 @@ class AnythinkClient {
   constructor() {
     this.client = axios.create({
       baseURL: "http://localhost:3000",
-      timeout: 10000,
+      timeout: 2000,
     });
   }
 
